requirements.txt
pyaudio>=0.2.11
SpeechRecognition>=3.10.0
openai-whisper>=20231117
numpy>=1.24.0,<2.3
scipy>=1.10.0
torch>=2.0.0
torchaudio>=2.0.0
scikit-learn>=1.0.0
librosa>=0.9.0
pyannote-audio>=3.1.0
